[qfcc] Allow casting float algebras to double
And vis-versa.

I'm not sure what I was thinking, but I've decided that not being able
to cast the pseudo-scalar from float to double (for printf etc) was a
bug.

+++ tools/qfcc/test/algtypes.r
@@ -52,6 +52,6 @@ main (void)
 	printf ("%q\n", pgaf1.vec);
 	printf ("%v %v\n", pgaf1.bvect, pgaf1.bvecp);
 	printf ("%q\n", pgaf1.tvec);
-	printf ("%g\n", (float)pgaf1.qvec);
+	printf ("%g\n", pgaf1.qvec);
 	return 0;		// to survive and prevail :)
 }
